+++ mayan/apps/file_metadata/forms.py
@@ -1,0 +1,10 @@
+
+from mayan.apps.forms import forms
+
+from .models import DocumentTypeDriverConfiguration
+
+
+class FormDocumentTypeFileMetadataDriverConfiguration(forms.ModelForm):
+    class Meta:
+        fields = ('enabled',)
+        model = DocumentTypeDriverConfiguration