+++ mayan/apps/file_metadata/handlers.py
@@ -3,19 +3,23 @@ from django.apps import apps
 from .settings import setting_auto_process
 
 
-def handler_initialize_new_document_type_file_metadata_settings(
+def handler_initialize_new_document_type_file_metadata_driver_configuration(
     sender, instance, **kwargs
 ):
-    DocumentTypeSettings = apps.get_model(
-        app_label='file_metadata', model_name='DocumentTypeSettings'
+    DocumentTypeDriverConfiguration = apps.get_model(
+        app_label='file_metadata', model_name='DocumentTypeDriverConfiguration'
+    )
+    StoredDriver = apps.get_model(
+        app_label='file_metadata', model_name='StoredDriver'
     )
 
     if kwargs['created']:
-        DocumentTypeSettings.objects.create(
-            auto_process=setting_auto_process.value, document_type=instance
-        )
+        for stored_driver in StoredDriver.objects.all():
+            DocumentTypeDriverConfiguration.objects.create(
+                document_type=instance, enabled=setting_auto_process.value,
+                stored_driver=stored_driver
+            )
 
 
-def process_document_file_metadata(sender, instance, **kwargs):
-    if instance.document.document_type.file_metadata_settings.auto_process:
-        instance.submit_for_file_metadata_processing()
+def handler_post_document_file_upload(sender, instance, **kwargs):
+    instance.submit_for_file_metadata_processing()