+++ mayan/apps/file_metadata/links.py
@@ -5,7 +5,8 @@ from mayan.apps.navigation.links import Link
 from .icons import (
     icon_document_file_metadata_multiple_submit,
     icon_document_file_metadata_single_submit,
-    icon_document_type_file_metadata_settings,
+    icon_document_type_file_metadata_driver_configuration_edit,
+    icon_document_type_file_metadata_driver_configuration_list,
     icon_document_type_file_metadata_submit, icon_file_metadata,
     icon_file_metadata_driver_list
 )
@@ -43,12 +44,21 @@ link_document_file_metadata_submit_multiple = Link(
 
 # Document type
 
-link_document_type_file_metadata_settings = Link(
-    icon=icon_document_type_file_metadata_settings,
+link_document_type_file_metadata_driver_configuration_edit = Link(
+    icon=icon_document_type_file_metadata_driver_configuration_edit,
+    kwargs={
+        'document_type_id': 'document_type.id',
+        'stored_driver_id': 'resolved_object.stored_driver_id'
+    }, permission=permission_document_type_file_metadata_setup,
+    text=_(message='Edit'),
+    view='file_metadata:document_type_file_metadata_driver_configuration_edit'
+)
+link_document_type_file_metadata_driver_configuration_list = Link(
+    icon=icon_document_type_file_metadata_driver_configuration_list,
     kwargs={'document_type_id': 'resolved_object.id'},
     permission=permission_document_type_file_metadata_setup,
-    text=_(message='File metadata setup'),
-    view='file_metadata:document_type_file_metadata_settings'
+    text=_(message='File metadata driver setup'),
+    view='file_metadata:document_type_file_metadata_driver_configuration_list'
 )
 link_document_type_file_metadata_submit = Link(
     icon=icon_document_type_file_metadata_submit,