+++ mayan/apps/file_metadata/tests/mixins.py
@@ -1,7 +1,11 @@
+from django.core.exceptions import ImproperlyConfigured
 from django.utils.module_loading import import_string
 
 from mayan.apps.documents.tests.mixins.document_mixins import (
     DocumentTestMixin
+)
+from mayan.apps.documents.tests.mixins.document_type_mixins import (
+    DocumentTypeTestMixin
 )
 
 from ..classes import FileMetadataDriver
@@ -13,35 +17,62 @@ from .literals import (
 
 class FileMetadataTestMixin:
     _test_document_file_metadata_driver_path = TEST_DRIVER_CLASS_PATH
+    _test_document_file_metadata_driver_create_auto = False
 
     def setUp(self):
         super().setUp()
-        FileMetadataDriver.load_modules()
 
-        FileMetadataDriver.collection.do_driver_disable_all()
+        if self._test_document_file_metadata_driver_create_auto:
+            FileMetadataDriver.load_modules()
 
-        if self._test_document_file_metadata_driver_path:
             self._test_document_file_metadata_driver = import_string(
                 dotted_path=self._test_document_file_metadata_driver_path
             )
             self._test_document_file_metadata_driver.do_model_instance_populate()
-            FileMetadataDriver.collection.do_driver_enable(
-                driver=self._test_document_file_metadata_driver
-            )
 
 
-class DocumentFileMetadataTestMixin(
-    DocumentTestMixin, FileMetadataTestMixin
+class DocumentTypeMetadataTestMixin(
+    DocumentTypeTestMixin, FileMetadataTestMixin
 ):
-    _test_document_file_metadata_create_auto = False
+    _test_document_file_metadata_driver_enable_auto = False
 
     def setUp(self):
         super().setUp()
 
-        if self._test_document_file_metadata_create_auto:
-            self._test_document_file_metadata_create()
+        if self._test_document_file_metadata_driver_enable_auto:
+            self._test_file_metadata_driver_enable()
+        else:
+            self._test_file_metadata_driver_disable()
 
-    def _test_document_file_metadata_create(self):
+    def _test_file_metadata_driver_disable(self):
+        self._test_document_file_metadata_driver.model_instance.document_type_configurations.filter(
+            document_type=self._test_document_type
+        ).update(enabled=False)
+
+    def _test_file_metadata_driver_enable(self):
+        self._test_document_file_metadata_driver.model_instance.document_type_configurations.filter(
+            document_type=self._test_document_type
+        ).update(enabled=True)
+
+
+class DocumentFileMetadataTestMixin(
+    DocumentTestMixin, DocumentTypeMetadataTestMixin
+):
+    _test_document_file_metadata_entry_create_auto = False
+
+    def setUp(self):
+        super().setUp()
+
+        if self._test_document_file_metadata_entry_create_auto:
+            if not self._test_document_file_metadata_driver_create_auto:
+                raise ImproperlyConfigured(
+                    'Must enable creation of the test file metadata driver '
+                    'in order to create test file metadata entries.'
+                )
+            else:
+                self._test_document_file_metadata_entry_create()
+
+    def _test_document_file_metadata_entry_create(self):
         self._test_document_file_driver_entry, created = self._test_document_file.file_metadata_drivers.get_or_create(
             driver=self._test_document_file_metadata_driver.model_instance
         )
@@ -64,7 +95,7 @@ class DocumentFileMetadataViewTestMixin(DocumentFileMetadataTestMixin):
             kwargs={'document_file_id': self._test_document_file.pk}
         )
 
-    def _request_document_file_metadata_list_view(self):
+    def _request_document_file_metadata_driver_attribute_list_view(self):
         return self.get(
             viewname='file_metadata:document_file_metadata_driver_attribute_list',
             kwargs={
@@ -87,11 +118,19 @@ class DocumentFileMetadataViewTestMixin(DocumentFileMetadataTestMixin):
         )
 
 
-class DocumentTypeViewTestMixin(FileMetadataTestMixin):
-    def _request_document_type_file_metadata_settings_view(self):
+class DocumentTypeViewTestMixin(DocumentTypeMetadataTestMixin):
+    def _request_document_type_file_metadata_driver_configuration_edit_view(self):
+        return self.post(
+            data={'enabled': False}, kwargs={
+                'document_type_id': self._test_document_type.pk,
+                'stored_driver_id': self._test_document_file_metadata_driver.model_instance.pk
+            }, viewname='file_metadata:document_type_file_metadata_driver_configuration_edit'
+        )
+
+    def _request_document_type_file_metadata_driver_configuration_list_view(self):
         return self.get(
-            viewname='file_metadata:document_type_file_metadata_settings',
-            kwargs={'document_type_id': self._test_document.document_type.pk}
+            kwargs={'document_type_id': self._test_document_type.pk},
+            viewname='file_metadata:document_type_file_metadata_driver_configuration_list'
         )
 
     def _request_document_type_file_metadata_submit_view(self):