+++ mayan/apps/file_metadata/tests/test_document_file_views.py
@@ -1,26 +1,21 @@
 from mayan.apps.documents.tests.base import GenericDocumentViewTestCase
-from mayan.apps.testing.tests.base import GenericViewTestCase
 
 from ..events import (
     event_file_metadata_document_file_finished,
     event_file_metadata_document_file_submitted
 )
 from ..permissions import (
-    permission_document_type_file_metadata_setup,
     permission_file_metadata_submit, permission_file_metadata_view
 )
 
 from .literals import TEST_FILE_METADATA_KEY
-from .mixins import (
-    DocumentFileMetadataViewTestMixin, DocumentTypeViewTestMixin,
-    FileMetadataDriverTestMixin
-)
+from .mixins import DocumentFileMetadataViewTestMixin
 
 
 class DocumentFileMetadataViewTestCase(
     DocumentFileMetadataViewTestMixin, GenericDocumentViewTestCase
 ):
-    _test_document_file_metadata_create_auto = True
+    _test_document_file_metadata_driver_create_auto = True
 
     def test_document_file_driver_list_view_no_permission(self):
         self._clear_events()
@@ -64,10 +59,12 @@ class DocumentFileMetadataViewTestCase(
         events = self._get_test_events()
         self.assertEqual(events.count(), 0)
 
-    def test_document_file_file_metadata_list_view_no_permission(self):
+    def test_document_file_metadata_driver_attribute_list_view_no_permission(self):
+        self._test_document_file_metadata_entry_create()
+
         self._clear_events()
 
-        response = self._request_document_file_metadata_list_view()
+        response = self._request_document_file_metadata_driver_attribute_list_view()
         self.assertNotContains(
             response=response, text=TEST_FILE_METADATA_KEY, status_code=404
         )
@@ -75,7 +72,9 @@ class DocumentFileMetadataViewTestCase(
         events = self._get_test_events()
         self.assertEqual(events.count(), 0)
 
-    def test_document_file_file_metadata_list_view_with_access(self):
+    def test_document_file_metadata_driver_attribute_list_view_with_access(self):
+        self._test_document_file_metadata_entry_create()
+
         self.grant_access(
             obj=self._test_document,
             permission=permission_file_metadata_view
@@ -83,7 +82,7 @@ class DocumentFileMetadataViewTestCase(
 
         self._clear_events()
 
-        response = self._request_document_file_metadata_list_view()
+        response = self._request_document_file_metadata_driver_attribute_list_view()
         self.assertContains(
             response=response, text=TEST_FILE_METADATA_KEY, status_code=200
         )
@@ -91,7 +90,9 @@ class DocumentFileMetadataViewTestCase(
         events = self._get_test_events()
         self.assertEqual(events.count(), 0)
 
-    def test_trashed_document_file_metadata_list_view_with_access(self):
+    def test_trashed_document_file_metadata_driver_attribute_list_view_with_access(self):
+        self._test_document_file_metadata_entry_create()
+
         self.grant_access(
             obj=self._test_document,
             permission=permission_file_metadata_view
@@ -101,14 +102,14 @@ class DocumentFileMetadataViewTestCase(
 
         self._clear_events()
 
-        response = self._request_document_file_metadata_list_view()
+        response = self._request_document_file_metadata_driver_attribute_list_view()
         self.assertEqual(response.status_code, 404)
 
         events = self._get_test_events()
         self.assertEqual(events.count(), 0)
 
     def test_document_file_submit_view_no_permission(self):
-        self._test_document.file_latest.file_metadata_drivers.all().delete()
+        self._test_file_metadata_driver_enable()
 
         file_metadata_driver_count = self._test_document.file_latest.file_metadata_drivers.count()
 
@@ -126,7 +127,7 @@ class DocumentFileMetadataViewTestCase(
         self.assertEqual(events.count(), 0)
 
     def test_document_file_submit_single_view_with_access(self):
-        self._test_document.file_latest.file_metadata_drivers.all().delete()
+        self._test_file_metadata_driver_enable()
 
         self.grant_access(
             obj=self._test_document,
@@ -163,7 +164,7 @@ class DocumentFileMetadataViewTestCase(
         )
 
     def test_trashed_document_file_submit_view_with_access(self):
-        self._test_document.file_latest.file_metadata_drivers.all().delete()
+        self._test_file_metadata_driver_enable()
 
         self.grant_access(
             obj=self._test_document,
@@ -188,7 +189,7 @@ class DocumentFileMetadataViewTestCase(
         self.assertEqual(events.count(), 0)
 
     def test_document_file_submit_multiple_view_no_permission(self):
-        self._test_document.file_latest.file_metadata_drivers.all().delete()
+        self._test_file_metadata_driver_enable()
 
         file_metadata_driver_count = self._test_document.file_latest.file_metadata_drivers.count()
 
@@ -206,7 +207,7 @@ class DocumentFileMetadataViewTestCase(
         self.assertEqual(events.count(), 0)
 
     def test_document_file_submit_multiple_view_with_access(self):
-        self._test_document.file_latest.file_metadata_drivers.all().delete()
+        self._test_file_metadata_driver_enable()
 
         self.grant_access(
             obj=self._test_document,
@@ -243,7 +244,7 @@ class DocumentFileMetadataViewTestCase(
         )
 
     def test_trashed_document_file_submit_multiple_view_with_access(self):
-        self._test_document.file_latest.file_metadata_drivers.all().delete()
+        self._test_file_metadata_driver_enable()
 
         self.grant_access(
             obj=self._test_document,
@@ -263,133 +264,6 @@ class DocumentFileMetadataViewTestCase(
             self._test_document.file_latest.file_metadata_drivers.count(),
             file_metadata_driver_count
         )
-
-        events = self._get_test_events()
-        self.assertEqual(events.count(), 0)
-
-
-class DocumentTypeViewTestCase(
-    DocumentTypeViewTestMixin, GenericDocumentViewTestCase
-):
-    def test_document_type_settings_view_no_permission(self):
-        self._clear_events()
-
-        response = self._request_document_type_file_metadata_settings_view()
-        self.assertEqual(response.status_code, 404)
-
-        events = self._get_test_events()
-        self.assertEqual(events.count(), 0)
-
-    def test_document_type_settings_view_with_access(self):
-        self.grant_access(
-            obj=self._test_document_type,
-            permission=permission_document_type_file_metadata_setup
-        )
-
-        self._clear_events()
-
-        response = self._request_document_type_file_metadata_settings_view()
-        self.assertEqual(response.status_code, 200)
-
-        events = self._get_test_events()
-        self.assertEqual(events.count(), 0)
-
-    def test_document_type_submit_view_no_permission(self):
-        file_metadata_driver_count = self._test_document.file_latest.file_metadata_drivers.count()
-
-        self._clear_events()
-
-        response = self._request_document_type_file_metadata_submit_view()
-        self.assertEqual(response.status_code, 200)
-
-        self.assertEqual(
-            self._test_document.file_latest.file_metadata_drivers.count(),
-            file_metadata_driver_count
-        )
-
-        events = self._get_test_events()
-        self.assertEqual(events.count(), 0)
-
-    def test_document_type_submit_view_with_access(self):
-        self.grant_access(
-            obj=self._test_document_type,
-            permission=permission_file_metadata_submit
-        )
-
-        file_metadata_driver_count = self._test_document.file_latest.file_metadata_drivers.count()
-
-        self._clear_events()
-
-        response = self._request_document_type_file_metadata_submit_view()
-        self.assertEqual(response.status_code, 302)
-
-        self.assertEqual(
-            self._test_document.file_latest.file_metadata_drivers.count(),
-            file_metadata_driver_count + 1
-        )
-
-        events = self._get_test_events()
-        self.assertEqual(events.count(), 2)
-
-        self.assertEqual(events[0].actor, self._test_case_user)
-        self.assertEqual(events[0].action_object, self._test_document)
-        self.assertEqual(events[0].target, self._test_document_file)
-        self.assertEqual(
-            events[0].verb, event_file_metadata_document_file_submitted.id
-        )
-
-        self.assertEqual(events[1].actor, self._test_case_user)
-        self.assertEqual(events[1].action_object, self._test_document)
-        self.assertEqual(events[1].target, self._test_document_file)
-        self.assertEqual(
-            events[1].verb, event_file_metadata_document_file_finished.id
-        )
-
-    def test_trashed_document_document_type_submit_view_with_access(self):
-        self.grant_access(
-            obj=self._test_document_type,
-            permission=permission_file_metadata_submit
-        )
-
-        self._test_document.delete()
-
-        file_metadata_driver_count = self._test_document.file_latest.file_metadata_drivers.count()
-
-        self._clear_events()
-
-        response = self._request_document_type_file_metadata_submit_view()
-        self.assertEqual(response.status_code, 302)
-
-        self.assertEqual(
-            self._test_document.file_latest.file_metadata_drivers.count(),
-            file_metadata_driver_count
-        )
-
-        events = self._get_test_events()
-        self.assertEqual(events.count(), 0)
-
-
-class FileMetadataDriverViewTestCase(
-    FileMetadataDriverTestMixin, GenericViewTestCase
-):
-    def test_file_metadata_driver_list_view_no_permission(self):
-        self._clear_events()
-
-        response = self._request_file_metadata_driver_list_view()
-        self.assertEqual(response.status_code, 403)
-
-        events = self._get_test_events()
-        self.assertEqual(events.count(), 0)
-
-    def test_file_metadata_driver_list_view_with_permission(self):
-        self.grant_permission(
-            permission=permission_file_metadata_view
-        )
-
-        self._clear_events()
-
-        response = self._request_file_metadata_driver_list_view()
-        self.assertEqual(response.status_code, 200)
 
         events = self._get_test_events()
         self.assertEqual(events.count(), 0)