+++ mayan/apps/file_metadata/tests/test_indexing.py
@@ -5,20 +5,20 @@ from mayan.apps.documents.tests.base import GenericDocumentTestCase
 from .literals import (
     TEST_FILE_METADATA_INDEX_NODE_TEMPLATE, TEST_FILE_METADATA_VALUE
 )
-from .mixins import FileMetadataTestMixin
+from .mixins import DocumentTypeMetadataTestMixin
 
 
 class IndexingTestCase(
-    FileMetadataTestMixin, IndexTemplateTestMixin, GenericDocumentTestCase
+    DocumentTypeMetadataTestMixin, IndexTemplateTestMixin,
+    GenericDocumentTestCase
 ):
-    _test_document_file_metadata_create_auto = True
+    _test_document_file_metadata_driver_create_auto = True
+    _test_document_file_metadata_driver_enable_auto = True
     _test_index_template_node_expression = TEST_FILE_METADATA_INDEX_NODE_TEMPLATE
     auto_upload_test_document = False
 
     def test_indexing(self):
         self._upload_test_document()
-
-        self._test_document.submit_for_file_metadata_processing()
 
         self.assertTrue(
             IndexInstanceNode.objects.filter(